Split the serialization logic from utils/object.ts; added the injected serialiser to the BlockItemStore
